Added new functionality
The program can now update existing information. It does this by having a index, where the infromation is stored. That information it can reuse, contra the one in the excel document.

It can now also take the books in a file tree. It now usees os.walk instead of os.serchdir.