New methods and a better method for choosing methods
The new method chooser was split. One for deciding what methods could be for that book and the other for testing what method works
It makes the program more servicable and modular, it makes it much easyer to make new methods.
There was alos some new methods added

Thewre was a warning system added, with that it gets mucth easyer to find issues and create new methods

In the readme file an issue was clarefyed